CC-34148 Fixes after CR
