+++ src/SprykerEco/Zed/AmazonQuicksight/Business/AmazonQuicksightFacadeInterface.php
@@ -36,9 +36,9 @@ interface AmazonQuicksightFacadeInterface
      * Specification:
      * - Expects `UserCollectionResponseTransfer.users.quicksightUser.role` to be set.
      * - Filters out users if `UserTransfer.quicksightUser.role` is not set.
-     * - Filters out users with already persisted quicksight user.
-     * - Filters out users with statuses not applicable for registering quicksight user.
-     * - Uses {@link \SprykerEco\Zed\AmazonQuicksight\AmazonQuicksightConfig::getUserStatusesApplicableForQuicksightUserRegistration()} to get a list of user statuses applicable for registering quicksight user.
+     * - Filters out users with already persisted Quicksight user.
+     * - Filters out users with statuses not applicable for registering a Quicksight user.
+     * - Uses {@link \SprykerEco\Zed\AmazonQuicksight\AmazonQuicksightConfig::getUserStatusesApplicableForQuicksightUserRegistration()} to get a list of user statuses applicable for registering a Quicksight user.
      * - Sends request to AWS API to register Quicksight users. For more information see {@link https://docs.aws.amazon.com/quicksight/latest/APIReference/API_RegisterUser.html}.
      * - Adds errors to `UserCollectionResponseTransfer.errors` if Quicksight user registration failed.
      * - Persists successfully registered Quicksight users in the database.
@@ -56,13 +56,12 @@ interface AmazonQuicksightFacadeInterface
 
     /**
      * Specification:
-     * - Filters out users with statuses not applicable for deleting quicksight user.
-     * - Uses {@link \SprykerEco\Zed\AmazonQuicksight\AmazonQuicksightConfig::getUserStatusesApplicableForQuicksightUserDeletion()} to get a list of user statuses applicable for deleting quicksight user.
-     * - Filters out users without persisted quicksight user.
+     * - Filters out users with statuses not applicable for deleting a Quicksight user.
+     * - Uses {@link \SprykerEco\Zed\AmazonQuicksight\AmazonQuicksightConfig::getUserStatusesApplicableForQuicksightUserDeletion()} to get a list of user statuses applicable for deleting a Quicksight user.
+     * - Filters out users without persisted Quicksight user.
      * - Sends request to AWS API to delete Quicksight users. For more information see {@link https://docs.aws.amazon.com/quicksight/latest/APIReference/API_DeleteUser.html}.
-     * - If AWS API call returns error it is added to `QuicksightUserCollectionResponseTransfer.errors` with `idQuicksightUser` as entity identifier.
-     * - If AWS API call returns error, quicksight user will not be deleted from persistence.
-     * - Deletes persistence quicksight users that were successfully deleted from Quicksight.
+     * - If the AWS API call returns an error, the Quicksight user will not be deleted from persistence.
+     * - Deletes from persistence Quicksight users that were successfully deleted from Quicksight.
      * - Adds errors to `UserCollectionResponseTransfer.errors` if any occurs.
      *
      * @api
@@ -99,34 +98,34 @@ interface AmazonQuicksightFacadeInterface
 
     /**
      * Specification:
-     * - Sends request to AWS API to get list of registered quicksight users. For more information see {@link https://docs.aws.amazon.com/quicksight/latest/APIReference/API_ListUsers.html}.
-     * - Filters out quicksight users with unsupported roles using {@link \SprykerEco\Zed\AmazonQuicksight\AmazonQuicksightConfig::getQuicksightUserRoles()}.
+     * - Sends request to AWS API to get list of registered Quicksight users. For more information see {@link https://docs.aws.amazon.com/quicksight/latest/APIReference/API_ListUsers.html}.
+     * - Filters out Quicksight users with unsupported roles using {@link \SprykerEco\Zed\AmazonQuicksight\AmazonQuicksightConfig::getQuicksightUserRoles()}.
      * - Fetches user transfers from persistence.
-     * - Filters out user transfers with persisted quicksight user.
-     * - Matches registered quicksight users with users by username.
-     * - Persists matched quicksight users.
+     * - Filters out user transfers with persisted Quicksight user.
+     * - Matches registered on Quicksight side Quicksight users with users from persistence by username.
+     * - Persists matched Quicksight users.
      * - Adds errors to `QuicksightUserCollectionResponseTransfer.errors` if any occurs.
-     * - Returns `QuicksightUserCollectionResponseTransfer` with persisted quicksight users and errors if any occurs.
+     * - Returns `QuicksightUserCollectionResponseTransfer` with persisted Quicksight users and errors if any occurs.
      *
      * @api
      *
      * @return \Generated\Shared\Transfer\QuicksightUserCollectionResponseTransfer
      */
-    public function createQuicksightUsersForRegisteredQuicksightUsersMatchedExistingUsers(): QuicksightUserCollectionResponseTransfer;
+    public function createMatchedQuicksightUsers(): QuicksightUserCollectionResponseTransfer;
 
     /**
      * Specification:
-     * - Sends request to AWS API to get list of registered quicksight users. For more information see {@link https://docs.aws.amazon.com/quicksight/latest/APIReference/API_ListUsers.html}.
-     * - Filters out quicksight users with unsupported roles using {@link \SprykerEco\Zed\AmazonQuicksight\AmazonQuicksightConfig::getQuicksightUserRoles()}.
+     * - Sends request to AWS API to get list of registered Quicksight users. For more information see {@link https://docs.aws.amazon.com/quicksight/latest/APIReference/API_ListUsers.html}.
+     * - Filters out Quicksight users with unsupported roles using {@link \SprykerEco\Zed\AmazonQuicksight\AmazonQuicksightConfig::getQuicksightUserRoles()}.
      * - Fetches user transfers from persistence.
-     * - Finds registered quicksight users not matched with persisted users.
-     * - Sends request to AWS API to delete quicksight users. For more information see {@link https://docs.aws.amazon.com/quicksight/latest/APIReference/API_DeleteUserByPrincipalId.html}.
+     * - Finds registered on Quicksight side Quicksight users not matched with users from persistence.
+     * - Sends request to AWS API to delete Quicksight users. For more information see {@link https://docs.aws.amazon.com/quicksight/latest/APIReference/API_DeleteUserByPrincipalId.html}.
      * - Adds errors to `QuicksightUserCollectionResponseTransfer.errors` if any occurs.
-     * - Returns `QuicksightUserCollectionResponseTransfer` with deleted quicksight users and errors if any occurs.
+     * - Returns `QuicksightUserCollectionResponseTransfer` with deleted Quicksight users and errors if any occurs.
      *
      * @api
      *
      * @return \Generated\Shared\Transfer\QuicksightUserCollectionResponseTransfer
      */
-    public function deleteRegisteredQuicksightUsersNotMatchedWithExistingUsers(): QuicksightUserCollectionResponseTransfer;
+    public function deleteNotMatchedQuicksightUsers(): QuicksightUserCollectionResponseTransfer;
 }
