+++ src/SprykerEco/Zed/AmazonQuicksight/Communication/Plugin/User/CreateQuicksightUserPostUpdatePlugin.php
@@ -20,6 +20,15 @@ class CreateQuicksightUserPostUpdatePlugin extends AbstractPlugin implements Use
 {
     /**
      * {@inheritDoc}
+     * - Expects `UserCollectionResponseTransfer.users.quicksightUser.role` to be set.
+     * - Filters out users if `UserTransfer.quicksightUser.role` is not set.
+     * - Filters out users with already persisted Quicksight user.
+     * - Filters out users with statuses not applicable for registering a Quicksight user.
+     * - Uses {@link \SprykerEco\Zed\AmazonQuicksight\AmazonQuicksightConfig::getUserStatusesApplicableForQuicksightUserRegistration()} to get a list of user statuses applicable for registering a Quicksight user.
+     * - Sends request to AWS API to register Quicksight users. For more information see {@link https://docs.aws.amazon.com/quicksight/latest/APIReference/API_RegisterUser.html}.
+     * - Adds errors to `UserCollectionResponseTransfer.errors` if Quicksight user registration failed.
+     * - Persists successfully registered Quicksight users in the database.
+     * - Returns `UserCollectionResponseTransfer` with updated `UserTransfers`.
      *
      * @api
      *
