+++ src/SprykerEco/Zed/AmazonQuicksight/Communication/Plugin/User/QuicksightUserPostCreatePlugin.php
@@ -21,8 +21,10 @@ class QuicksightUserPostCreatePlugin extends AbstractPlugin implements UserPostC
     /**
      * {@inheritDoc}
      * - Expects `UserCollectionResponseTransfer.users.quicksightUser.role` to be set.
-     * - Does nothing if `UserTransfer.quicksightUser.role` is not set.
-     * - Filters out users with already persisted quicksight user.
+     * - Filters out users if `UserTransfer.quicksightUser.role` is not set.
+     * - Filters out users with already persisted Quicksight user.
+     * - Filters out users with statuses not applicable for registering a Quicksight user.
+     * - Uses {@link \SprykerEco\Zed\AmazonQuicksight\AmazonQuicksightConfig::getUserStatusesApplicableForQuicksightUserRegistration()} to get a list of user statuses applicable for registering a Quicksight user.
      * - Sends request to AWS API to register Quicksight users. For more information see {@link https://docs.aws.amazon.com/quicksight/latest/APIReference/API_RegisterUser.html}.
      * - Adds errors to `UserCollectionResponseTransfer.errors` if Quicksight user registration failed.
      * - Persists successfully registered Quicksight users in the database.
